Удаление комментариев по средствам аякс

+++ protected/controllers/SiteController.php
@@ -66,11 +66,10 @@ class SiteController extends Controller
 
     public function actionRemovecomment($id)
     {
-
-        // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-//        if (!isset($_GET['ajax']))
-//            $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
-        return Comment::model()->deleteByPk($id);
+        if (Yii::app()->user->checkAccess(User::ROLE_ADMIN))
+            return Comment::model()->deleteByPk($id);
+        else
+            return false;
     }
 
     public function actionLogin()
